@@ -7,19 +7,19 @@ import { getAccessToken } from '../../utils/auth'
 
 const { COMIC, FAVOURITISE, RATE, BOOKMARK } = COMIC_QUERY_KEYS
 
-export const rateComic = async ({ slug, request }: { slug: string; request: RateComic }): Promise<void> => {
-  const accessToken = getAccessToken()
-  await fetchWrapper<void>({
+export const rateComic = async ({ slug, request }: { slug: string; request: RateComic }): Promise<string> => {
+  const accessToken = await getAccessToken()
+  const response = await fetchWrapper({
     accessToken,
     path: `${COMIC}/${RATE}/${slug}`,
     body: request,
     method: 'PATCH',
-    isTextResponse: true,
   })
+  return response.errorMessage ?? ''
 }
 
 export const favouritiseComic = async (slug: string): Promise<void> => {
-  const accessToken = getAccessToken()
+  const accessToken = await getAccessToken()
   await fetchWrapper<void>({
     accessToken,
     path: `${COMIC}/${FAVOURITISE}/${slug}`,
@@ -29,7 +29,7 @@ export const favouritiseComic = async (slug: string): Promise<void> => {
 }
 
 export const bookmarkComic = async (slug: string): Promise<void> => {
-  const accessToken = getAccessToken()
+  const accessToken = await getAccessToken()
   await fetchWrapper<void>({
     accessToken,
     path: `${COMIC}/${BOOKMARK}/${slug}`,
